refactor: move built-in plugins to top-level plugins/ dir
